[DEV-1591] Fixed build, refactoring
