+++ src/subdomains/user/enums/kyc.enum.ts
