+++ src/subdomains/user/services/kyc-sync.service.ts
@@ -8,8 +8,8 @@ import { DfxLogger } from 'src/shared/services/dfx-logger';
 import { Lock } from 'src/shared/utils/lock';
 import { Util } from 'src/shared/utils/util';
 import { SettingService } from 'src/subdomains/master-data/setting/setting.service';
-import { KycStepName, KycStepStatus } from '../entities/kyc.enum';
 import { User } from '../entities/user.entity';
+import { KycStepName, KycStepStatus } from '../enums/kyc.enum';
 import { KycService } from './kyc.service';
 import { MandatorService } from './mandator.service';
 import { UserService } from './user.service';
