+++ src/subdomains/user/services/kyc.service.ts
@@ -6,9 +6,9 @@ import { UserService } from 'src/subdomains/user/services/user.service';
 import { KycDataDto } from '../api/dto/user-in.dto';
 import { UserInfoDto } from '../api/dto/user-out.dto';
 import { KycStep } from '../entities/kyc-step.entity';
-import { KycStepName, KycStepStatus } from '../entities/kyc.enum';
 import { User } from '../entities/user.entity';
-import { AccountType } from '../entities/user.enum';
+import { KycStepName, KycStepStatus } from '../enums/kyc.enum';
+import { AccountType } from '../enums/user.enum';
 
 export class KycService {
   private readonly logger = new DfxLogger(KycService);
