+++ src/subdomains/user/services/user.service.ts
@@ -6,7 +6,7 @@ import { SettingsDto } from '../api/dto/user-in.dto';
 import { UserInfoMapper } from '../api/dto/user-info.mapper';
 import { UserInfoDto } from '../api/dto/user-out.dto';
 import { User } from '../entities/user.entity';
-import { KycStatus } from '../entities/user.enum';
+import { KycStatus } from '../enums/user.enum';
 import { UserRepository } from '../repositories/user.repository';
 import { MandatorService } from './mandator.service';
